[FE] 전체적인 이름 수정 및 디렉토리 수정 (#45) (#46)
style: 전체적인 이름 수정 및 디렉토리 수정 #45
