+++ fe/src/pages/Issues.tsx
@@ -1,12 +1,15 @@
-import { useEffect } from "react";
-import {
-  IssueTableHeader,
-  // ChangeStateHeader,
-} from "components/Header/IssueTableHeader";
+import { useState, useEffect } from "react";
+import { IssueTableHeader } from "components/Table/IssueTableHeader";
 import { IssueListHeader } from "components/Header/IssueListHeader";
-import { IssuesTable } from "components/Table/IssuesTable";
+import { IssueTableList } from "components/Table/IssueTableList";
 
 export const IssuesPage = () => {
+  const [isChecked, setIsChecked] = useState(false);
+
+  const handleClickCheckBox = () => {
+    setIsChecked(!isChecked);
+  };
+
   useEffect(() => {
     fetch("http://3.39.194.109:8080/api/issues/open")
       .then((response) => response.json())
@@ -18,9 +21,11 @@ export const IssuesPage = () => {
   return (
     <>
       <IssueListHeader />
-      <IssueTableHeader />
-      {/* <ChangeStateHeader /> */}
-      <IssuesTable />
+      <IssueTableHeader
+        isChecked={isChecked}
+        onClickCheckBox={handleClickCheckBox}
+      />
+      <IssueTableList />
     </>
   );
 };
